third_party/WebKit/LayoutTests/fast/css/variables/resolve-duplicate-variables-in-style-attribute.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id='testElem'></div>
<script>
test(function() {
  testElem.setAttribute('style', '--foo:one;--foo:two;');
  assert_equals(testElem.style.length, 1);
  assert_equals(testElem.style.getPropertyValue('--foo'), 'two');
}, "the cascade applies correctly to variables in inline style.")
</script>
